refactor schemas using typescript
