+++ packages/schemas/package.json
@@ -8,10 +8,16 @@
   "keywords": [
     "sockethub",
     "schema",
-    "activity",
-    "streams"
+    "activity streams",
+    "activity objects",
+    "platforms"
   ],
-  "main": "./index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist/",
+    "src/"
+  ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/sockethub/sockethub.git",
@@ -19,17 +25,31 @@
   },
   "homepage": "https://github.com/sockethub/sockethub/tree/master/packages/schemas",
   "devDependencies": {
+    "@types/chai": "4.3.0",
+    "@types/debug": "4.1.7",
+    "@types/eslint": "8.4.1",
+    "@types/mocha": "9.1.0",
+    "@types/node": "17.0.13",
+    "@typescript-eslint/parser": "5.12.1",
     "c8": "7.11.0",
     "chai": "4.3.6",
-    "mocha": "9.2.1"
+    "eslint": "8.9.0",
+    "eslint-cli": "1.1.1",
+    "mocha": "9.2.1",
+    "ts-node": "10.5.0",
+    "typescript": "4.5.5"
   },
   "scripts": {
-    "clean": "npx rimraf coverage",
+    "build": "npx rimraf dist && tsc && ./scripts/export-json-schema.js",
+    "clean": "npx rimraf coverage dist",
     "clean:deps": "npx rimraf node_modules",
-    "compliance": "yarn run test && yarn run coverage",
-    "test": "yarn run build && c8 mocha **/*.test.js",
+    "compliance": "yarn test && yarn coverage",
     "coverage": "c8 check-coverage --statements 95 --branches 95 --functions 95 --lines 95",
-    "build": "./bin/build.js"
+    "lint": "eslint \"**/*.ts\"",
+    "lint:fix": "eslint --fix \"**/*.ts\"",
+    "test": "yarn build && yarn test:ts && yarn test:js",
+    "test:ts": "c8 mocha -r ts-node/register src/*.test.ts",
+    "test:js": "c8 mocha **/*.test.js"
   },
   "dependencies": {
     "ajv": "8.10.0",
