+++ packages/schemas/scripts/export-json-schema.js
@@ -1,0 +1,21 @@
+#!/usr/bin/env node
+const Ajv = require('ajv');
+const fs = require('fs');
+
+const ajv = new Ajv();
+
+const schemas = [
+  'activity-stream',
+  'activity-object',
+  'platform'
+];
+
+fs.mkdirSync('./dist/schemas/json');
+
+for (let fileName of schemas) {
+  // eslint-disable-next-line security-node/detect-non-literal-require-calls
+  const schema = require(`../dist/schemas/${fileName}`).default;
+  ajv.addSchema(schema, schema.id);
+  fd = fs.openSync(`./dist/schemas/json/${fileName}.json`, 'w+');
+  fs.writeSync(fd, JSON.stringify(schema, null, "\t"));
+}
