+++ packages/schemas/src/schemas/platform.ts
@@ -18,7 +18,7 @@
 
 // this schema defines the general structure of the schema object which should
 // be returned from platforms.
-module.exports = {
+export default {
   'type' : 'object',
   'required': [ 'version', 'messages' ],
   'additionalProperties': false,
